test: improve tests
